requirements.txt
dateparser==1.2.0
Flask==3.1.0
google_api_python_client==2.151.0
google_auth_oauthlib==0.4.6
openai==0.28.0
protobuf==5.29.2
python-dotenv==1.0.1
python_docx==1.1.2
Requests==2.32.3
requests_toolbelt==1.0.0
upstash_redis==1.2.0
